Break BlueprintId out into its own module
